requirements.txt
async-generator==1.10
attrs==20.3.0
beautifulsoup4==4.9.3
bleach==3.2.1
boto3==1.17.62
botocore==1.20.62
category-encoders==2.3.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.3
cryptography==36.0.0
dnspython==2.1.0
docutils==0.16
Flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.1.0
keyring==21.4.0
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
outcome==1.1.0
packaging==20.4
pandas==1.3.4
patsy==0.5.2
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2-binary==2.9.2
py==1.10.0
py-openaq==1.1.0
pycparser==2.21
Pygments==2.7.1
pymongo==3.12.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.2
pytest-flask==1.2.0
pytest-json-report==1.4.1
pytest-metadata==1.11.0
python-dateutil==2.8.1
pytz==2021.3
readme-renderer==26.0
requests==2.25.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==1.0.1
scipy==1.7.3
selenium==4.1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
statsmodels==0.13.1
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.50.1
trio==0.19.0
trio-websocket==0.9.2
twine==3.2.0
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==2.0.1
wsproto==1.0.0
